requirements.txt
opencv_python==4.4.0.44
matplotlib==3.3.2
pandas==1.0.1
numpy==1.18.1
torch==1.7.0
scikit_image==0.17.2
Pillow==8.0.1
scikit_learn==0.23.2
skimage==0.0
torchvision==0.8.2
